+++ build.gradle
@@ -1,8 +1,8 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
 
-  ext.KOTLIN_VERSION = "1.1.4-3"
-  ext.ANDROID_PLUGIN_VERSION = "3.0.0-beta5"
+  ext.KOTLIN_VERSION = "1.1.50"
+  ext.ANDROID_PLUGIN_VERSION = "3.0.0-beta6"
 
   repositories {
     google()
