--- epoxy-integrationtest/src/main/java/com/airbnb/epoxy/ModelWithClickListener.java
@@ -1,18 +1,0 @@
-package com.airbnb.epoxy;
-
-import android.view.View;
-
-class ModelWithClickListener extends EpoxyModel<View> {
-
-  @EpoxyAttribute View.OnClickListener clickListener;
-
-  @Override
-  protected int getDefaultLayout() {
-    return R.layout.model_with_click_listener;
-  }
-
-  @Override
-  public void bind(View view) {
-    view.setOnClickListener(clickListener);
-  }
-}
