--- epoxy-integrationtest/src/main/java/com/airbnb/epoxy/TestDataBindingModel.java
@@ -1,6 +1,0 @@
-package com.airbnb.epoxy;
-
-@EpoxyModelClass(layout = R.layout.model_with_data_binding)
-abstract class TestDataBindingModel extends DataBindingEpoxyModel {
-  @EpoxyAttribute String stringValue;
-}
