+++ epoxy-integrationtest/src/main/java/com/airbnb/epoxy/integrationtest/AdapterWithFieldAssigned.java
@@ -1,0 +1,14 @@
+package com.airbnb.epoxy.integrationtest;
+
+import com.airbnb.epoxy.AutoModel;
+import com.airbnb.epoxy.EpoxyController;
+
+public class AdapterWithFieldAssigned extends EpoxyController {
+
+  @AutoModel Model_ model1 = new Model_();
+
+  @Override
+  protected void buildModels() {
+    add(model1);
+  }
+}
