+++ epoxy-integrationtest/src/main/java/com/airbnb/epoxy/integrationtest/Model.java
@@ -1,10 +1,14 @@
-package com.airbnb.epoxy;
+package com.airbnb.epoxy.integrationtest;
 
 import android.widget.TextView;
 
+import com.airbnb.epoxy.EpoxyAttribute;
+import com.airbnb.epoxy.EpoxyModel;
+import com.airbnb.epoxy.EpoxyModelClass;
+
 @EpoxyModelClass
 public abstract class Model extends EpoxyModel<TextView> {
-  @EpoxyAttribute int value;
+  @EpoxyAttribute public int value;
 
   @Override
   protected int getDefaultLayout() {
