+++ epoxy-integrationtest/src/main/java/com/airbnb/epoxy/integrationtest/ViewWithAnnotationsForIntegrationTest.java
@@ -1,11 +1,13 @@
-package com.airbnb.epoxy;
+package com.airbnb.epoxy.integrationtest;
 
 import android.content.Context;
 import android.support.annotation.Nullable;
 import android.util.AttributeSet;
 import android.view.View;
 
+import com.airbnb.epoxy.ModelProp;
 import com.airbnb.epoxy.ModelProp.Option;
+import com.airbnb.epoxy.ModelView;
 
 @ModelView(defaultLayout = R.layout.view_with_annotations_for_integration_test)
 public class ViewWithAnnotationsForIntegrationTest extends View {
