+++ epoxy-integrationtest/src/main/java/com/airbnb/epoxy/integrationtest/autoaddautomodels/package-info.java
@@ -1,4 +1,4 @@
 @PackageEpoxyConfig(implicitlyAddAutoModels = true)
-package com.airbnb.epoxy.autoaddautomodels;
+package com.airbnb.epoxy.integrationtest.autoaddautomodels;
 
 import com.airbnb.epoxy.PackageEpoxyConfig;
