+++ epoxy-integrationtest/src/main/java/com/airbnb/epoxy/integrationtest/package-info.java
@@ -1,0 +1,4 @@
+@EpoxyDataBindingLayouts({R.layout.model_with_data_binding})
+package com.airbnb.epoxy.integrationtest;
+
+import com.airbnb.epoxy.EpoxyDataBindingLayouts;
