--- epoxy-integrationtest/src/main/java/com/airbnb/epoxy/package-info.java
@@ -1,2 +1,0 @@
-@EpoxyDataBindingLayouts({R.layout.model_with_data_binding})
-package com.airbnb.epoxy;
