+++ epoxy-integrationtest/src/main/res/layout/view_with_annotations_for_integration_test.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.airbnb.epoxy.ViewWithAnnotationsForIntegrationTest xmlns:android="http://schemas.android.com/apk/res/android"
+<com.airbnb.epoxy.integrationtest.ViewWithAnnotationsForIntegrationTest xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical" />
