+++ epoxy-integrationtest/src/test/java/com/airbnb/epoxy/BindModelIntegrationTest.java
@@ -2,6 +2,9 @@ package com.airbnb.epoxy;
 
 import android.widget.TextView;
 
+import com.airbnb.epoxy.integrationtest.BuildConfig;
+import com.airbnb.epoxy.integrationtest.Model_;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
