+++ epoxy-integrationtest/src/test/java/com/airbnb/epoxy/DataBindingModelIntegrationTest.java
@@ -6,6 +6,8 @@ import android.view.View.OnClickListener;
 import android.widget.Button;
 
 import com.airbnb.epoxy.DataBindingEpoxyModel.DataBindingHolder;
+import com.airbnb.epoxy.integrationtest.BuildConfig;
+import com.airbnb.epoxy.integrationtest.ModelWithDataBindingBindingModel_;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -30,7 +32,7 @@ public class DataBindingModelIntegrationTest {
   @Test
   public void createDataBindingModel() {
     SimpleEpoxyController controller = new SimpleEpoxyController();
-    TestDataBindingModel_ firstModel = new TestDataBindingModel_()
+    ModelWithDataBindingBindingModel_ firstModel = new ModelWithDataBindingBindingModel_()
         .stringValue("hello")
         .id(1);
 
@@ -46,7 +48,7 @@ public class DataBindingModelIntegrationTest {
     // Check that the requiredText was set on the view
     assertEquals(firstModel.stringValue(), ((Button) viewHolder.itemView).getText());
 
-    TestDataBindingModel_ secondModel = new TestDataBindingModel_()
+    ModelWithDataBindingBindingModel_ secondModel = new ModelWithDataBindingBindingModel_()
         .stringValue("hello again")
         .id(1);
