+++ epoxy-integrationtest/src/test/java/com/airbnb/epoxy/EpoxyAdapterIntegrationTest.java
@@ -1,5 +1,9 @@
 package com.airbnb.epoxy;
 
+import com.airbnb.epoxy.integrationtest.BuildConfig;
+import com.airbnb.epoxy.integrationtest.Model;
+import com.airbnb.epoxy.integrationtest.Model_;
+
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
