+++ epoxy-integrationtest/src/test/java/com/airbnb/epoxy/EpoxyModelValidationTest.java
@@ -1,8 +1,12 @@
 package com.airbnb.epoxy;
 
-import android.view.View;
-
 import com.airbnb.epoxy.EpoxyController.Interceptor;
+import com.airbnb.epoxy.integrationtest.BuildConfig;
+import com.airbnb.epoxy.integrationtest.ControllerWithAutoModel;
+import com.airbnb.epoxy.integrationtest.Model;
+import com.airbnb.epoxy.integrationtest.ModelChangesDuringBind_;
+import com.airbnb.epoxy.integrationtest.Model_;
+import com.airbnb.epoxy.integrationtest.R;
 
 import org.junit.Rule;
 import org.junit.Test;
