+++ epoxy-integrationtest/src/test/java/com/airbnb/epoxy/ModelClickListenerTest.java
@@ -5,6 +5,10 @@ import android.view.View;
 import android.view.View.OnClickListener;
 import android.widget.FrameLayout;
 
+import com.airbnb.epoxy.integrationtest.BuildConfig;
+import com.airbnb.epoxy.integrationtest.ModelWithClickListener_;
+import com.airbnb.epoxy.integrationtest.ModelWithLongClickListener_;
+
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
