+++ epoxy-integrationtest/src/test/java/com/airbnb/epoxy/OnModelBindListenerTest.java
@@ -3,6 +3,9 @@ package com.airbnb.epoxy;
 import android.support.v7.widget.RecyclerView.AdapterDataObserver;
 import android.view.View;
 
+import com.airbnb.epoxy.integrationtest.BuildConfig;
+import com.airbnb.epoxy.integrationtest.ModelWithClickListener_;
+
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
