+++ epoxy-processor/src/main/java/com/airbnb/epoxy/KotlinModelBuilderExtensionWriter.kt
@@ -78,7 +78,6 @@ internal class KotlinModelBuilderExtensionWriter(
             return build()
         }
 
-
     }
 
     private fun getMethodName(model: GeneratedModelInfo)
@@ -92,58 +91,6 @@ internal class KotlinModelBuilderExtensionWriter(
             .removeSuffix("Epoxy")
 }
 
-typealias JavaClassName = com.squareup.javapoet.ClassName
-typealias JavaTypeName = com.squareup.javapoet.TypeName
-typealias JavaWildcardTypeName = com.squareup.javapoet.WildcardTypeName
-typealias JavaArrayTypeName = com.squareup.javapoet.ArrayTypeName
-typealias JavaTypeVariableName = com.squareup.javapoet.TypeVariableName
-typealias JavaParametrizedTypeName = com.squareup.javapoet.ParameterizedTypeName
 
-typealias KotlinClassName = com.squareup.kotlinpoet.ClassName
-typealias KotlinParameterizedTypeName = com.squareup.kotlinpoet.ParameterizedTypeName
-typealias KotlinTypeName = com.squareup.kotlinpoet.TypeName
-typealias KotlinWildcardTypeName = com.squareup.kotlinpoet.WildcardTypeName
-typealias KotlinTypeVariableName = com.squareup.kotlinpoet.TypeVariableName
-
-fun JavaClassName.toKPoet() = KotlinClassName(
-        packageName(),
-        simpleName())
-
-// Does not support transferring annotations
-fun JavaWildcardTypeName.toKPoet() =
-        if (!lowerBounds.isEmpty()) {
-            KotlinWildcardTypeName.supertypeOf(lowerBounds.first().toKPoet())
-        } else {
-            KotlinWildcardTypeName.subtypeOf(upperBounds.first().toKPoet())
-        }
-
-// Does not support transferring annotations
-fun JavaParametrizedTypeName.toKPoet()
-        = KotlinParameterizedTypeName.get(
-        this.rawType.toKPoet(),
-        *typeArguments.toKPoet().toTypedArray())
-
-// Does not support transferring annotations
-fun JavaArrayTypeName.toKPoet()
-        = KotlinParameterizedTypeName.get(
-        KotlinClassName.bestGuess("kotlin.Array"),
-        this.componentType.toKPoet())
-
-// Does not support transferring annotations
-fun JavaTypeVariableName.toKPoet()
-        = KotlinTypeVariableName.invoke(
-        name,
-        *bounds.toKPoet().toTypedArray())
-
-fun JavaTypeName.toKPoet(): KotlinTypeName = when (this) {
-    is JavaClassName -> toKPoet()
-    is JavaParametrizedTypeName -> toKPoet()
-    is JavaArrayTypeName -> toKPoet()
-    is JavaTypeVariableName -> toKPoet()
-    is JavaWildcardTypeName -> toKPoet()
-    else -> throw IllegalArgumentException("Unsupported type: ${this::class.simpleName}")
-}
-
-fun <T : JavaTypeName> Iterable<T>.toKPoet() = map { it.toKPoet() }
